requirements.txt
discord.py==1.7.3
requests==2.26.0
websockets==9.1
web3==5.24.0